+++ orm/src/main/java/org/hibernate/tool/hbm2x/HibernateConfigurationExporter.java
@@ -20,7 +20,7 @@ import java.util.TreeMap;
 import org.hibernate.cfg.Environment;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.RootClass;
-import org.hibernate.tool.internal.export.AbstractExporter;
+import org.hibernate.tool.internal.export.common.AbstractExporter;
 
 /**
  * @author max
